+++ client/src/components/Article/ArticleDetail/ArticleDetailBack.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 import { SlArrowLeft } from "react-icons/sl";
