+++ client/src/components/Article/ArticleDetail/articledetailcontent/ArticleDetailBody.tsx
@@ -24,6 +24,7 @@ function ArticleDetailBody({ content }: props) {
 export default ArticleDetailBody;
 
 const MainContainer = styled.div`
+  display: flex;
   color: var(--color-main);
   width: 100%;
   padding: var(--large) var(--2x-large) var(--2x-large) var(--2x-large);
@@ -33,5 +34,18 @@ const MainContainer = styled.div`
 
   ul {
     padding-left: var(--large);
+  }
+
+  @media only screen and (max-width: 450px) {
+    padding: var(--large); 
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+    white-space: pre-line;
+    width: 100%;
+
+    img {
+      width: 80%;
+    }
   }
 `;
