+++ client/src/components/Article/ArticleDetail/articledetailmenu/ArticleDetailMenuItem.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import styled from 'styled-components';
-import Card from '../../UI/Card';
+import Card from '../../../UI/Card';
 
 interface MenuItemProps {
   title: string;
